Objects + reset text
The JFrame creates a new object depending on the Combo Box selection and the bufferedReader writes the ToString. The main issue here is that now, the reader does not separate entries by a new line. Also, we need to have a separate method for the code that resets the text boxes to reduce repetitiveness.